ERRORS FIXED

+++ MiniProject/MiniProject/Controllers/CourseGroupController.cs
@@ -1,6 +1,7 @@
 ﻿using Domain.Models;
 using Service.Helper;
 using Service.Services.Implementations;
+using Service.Services.Interfaces;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -27,22 +28,44 @@ namespace MiniProject.Controllers
         public void UpdateCourseGroup(CourseGroupService groupService)
         {
             int id = Helper.ReadValidatedInt("Enter ID to update:");
-            Console.Write("Enter new name: ");
-            string newName = Console.ReadLine();
-            Console.Write("Enter new teacher name: ");
-            string newTeacher = Console.ReadLine();
-            Console.Write("Enter new room: ");
-            string newRoom = Console.ReadLine();
 
-            groupService.Update(id, new CourseGroup { Name = newName, TeacherName = newTeacher, Room = newRoom });
-            Helper.ColorWrite(ConsoleColor.Green, "Group updated successfully!");
+            var group = groupService.GetById(id);
+
+            if (group == null)
+            {
+                Helper.ColorWrite(ConsoleColor.Red, "There is no group with given ID!");
+            }
+            else
+            {
+                Console.Write("Enter new name: ");
+                string newName = Console.ReadLine();
+                Console.Write("Enter new teacher name: ");
+                string newTeacher = Console.ReadLine();
+                Console.Write("Enter new room: ");
+                string newRoom = Console.ReadLine();
+
+
+                groupService.Update(id, new CourseGroup { Name = newName, TeacherName = newTeacher, Room = newRoom });
+                Helper.ColorWrite(ConsoleColor.Green, "Group updated successfully!");
+            }
         }
 
         public void DeleteCourseGroup(CourseGroupService groupService)
         {
             int id = Helper.ReadValidatedInt("Enter ID to delete:");
-            groupService.Delete(id);
-            Helper.ColorWrite(ConsoleColor.Green, "Group deleted successfully!");
+
+            var group = groupService.GetById(id);
+            if (group != null)
+            {
+
+                groupService.Delete(id);
+                Helper.ColorWrite(ConsoleColor.Green, "Group deleted successfully!");
+            }
+            else
+            {
+                Helper.ColorWrite(ConsoleColor.Red, "There is no group with given ID!");
+            }
+
         }
 
         public void GetCourseGroupById(CourseGroupService groupService)
@@ -58,8 +81,16 @@ namespace MiniProject.Controllers
         public void GetAllCourseGroups(CourseGroupService groupService)
         {
             var all = groupService.GetAll();
-            foreach (var g in all)
-                Console.WriteLine($"{g.Id}. {g.Name} | Teacher: {g.TeacherName} | Room: {g.Room}");
+
+            if (all.Count == 0)
+            {
+                Helper.ColorWrite(ConsoleColor.Red, "There is no group!");
+            }
+            else
+            {
+                foreach (var g in all)
+                    Console.WriteLine($"{g.Id}. {g.Name} | Teacher: {g.TeacherName} | Room: {g.Room}");
+            }
         }
 
         public void GetCourseGroupsByTeacher(CourseGroupService groupService)
