+++ MiniProject/MiniProject/Controllers/StudentController.cs
@@ -3,6 +3,7 @@ using Service.Helper;
 using Service.Services.Implementations;
 using System;
 using System.Collections.Generic;
+using System.ComponentModel.Design;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -25,21 +26,33 @@ namespace MiniProject.Controllers
         public void UpdateStudent(StudentService studentService)
         {
             int id = Helper.ReadValidatedInt("Enter student ID to update:");
-            Console.Write("Enter new name: ");
-            string newName = Console.ReadLine();
-            Console.Write("Enter new surname: ");
-            string newSurname = Console.ReadLine();
-            int newAge = Helper.ReadValidatedInt("Enter new age:");
 
-            studentService.Update(id, new Student { Name = newName, Surname = newSurname, Age = newAge });
-            Helper.ColorWrite(ConsoleColor.Green, "Student updated successfully!");
+            var student = studentService.GetById(id);
+
+            if (student != null)
+            {
+
+                Console.Write("Enter new name: ");
+                string newName = Console.ReadLine();
+                Console.Write("Enter new surname: ");
+                string newSurname = Console.ReadLine();
+                int newAge = Helper.ReadValidatedInt("Enter new age:");
+
+                studentService.Update(id, new Student { Name = newName, Surname = newSurname, Age = newAge });
+                Helper.ColorWrite(ConsoleColor.Green, "Student updated successfully!");
+            }
         }
 
         public void DeleteStudent(StudentService studentService)
         {
             int id = Helper.ReadValidatedInt("Enter student ID to delete:");
-            studentService.Delete(id);
-            Helper.ColorWrite(ConsoleColor.Green, "Student deleted successfully!");
+            var student = studentService.GetById(id);
+            if (student != null)
+            {
+                studentService.Delete(id);
+
+                Helper.ColorWrite(ConsoleColor.Green, "Student deleted successfully!");
+            }
         }
 
         public void GetStudentById(StudentService studentService)
@@ -48,31 +61,52 @@ namespace MiniProject.Controllers
             var student = studentService.GetById(id);
             if (student != null)
                 Console.WriteLine($"{student.Id}. {student.Name} {student.Surname} | Age: {student.Age} | Group: {student.CourseGroup?.Name}");
-            else
-                Helper.ColorWrite(ConsoleColor.Red, "Student not found!");
         }
 
         public void GetAllStudents(StudentService studentService)
         {
             var list = studentService.GetAll();
-            foreach (var s in list)
-                Console.WriteLine($"{s.Id}. {s.Name} {s.Surname} | {s.Age} | Group: {s.CourseGroup?.Name}");
+            if (list.Count == 0)
+            {
+                Helper.ColorWrite(ConsoleColor.Red, "There is no student!");
+            }
+            else
+            {
+                foreach (var s in list)
+                    Console.WriteLine($"{s.Id}. {s.Name} {s.Surname} | {s.Age} | Group: {s.CourseGroup?.Name}");
+
+            }
         }
 
         public void GetStudentsByAge(StudentService studentService)
         {
             int age = Helper.ReadValidatedInt("Enter age:");
             var list = studentService.GetAllByAge(age);
-            foreach (var s in list)
-                Console.WriteLine($"{s.Id}. {s.Name} {s.Surname} | Age: {s.Age}");
+            if (list.Count == 0)
+            {
+                Helper.ColorWrite(ConsoleColor.Red, "There is no student with given age!");
+            }
+            else
+            {
+                foreach (var s in list)
+                    Console.WriteLine($"{s.Id}. {s.Name} {s.Surname} | Age: {s.Age}");
+
+            }
         }
 
         public void GetStudentsByGroup(StudentService studentService)
         {
             int groupId = Helper.ReadValidatedInt("Enter group ID:");
             var list = studentService.GetAllByGroupId(groupId);
-            foreach (var s in list)
-                Console.WriteLine($"{s.Id}. {s.Name} {s.Surname}");
+            if (list.Count == 0)
+            {
+                Helper.ColorWrite(ConsoleColor.Red, "There is no student with given group ID!");
+            }
+            else
+            {
+                foreach (var s in list)
+                    Console.WriteLine($"{s.Id}. {s.Name} {s.Surname}");
+            }
         }
 
         public void SearchStudents(StudentService studentService)
